activity search and filter

+++ app/src/main/java/com/code4piter/blueskythinking/megapp/ui/activity/CamerasActivity.java
@@ -29,70 +29,72 @@ import retrofit2.Response;
 import retrofit2.Retrofit;
 
 public class CamerasActivity extends AppCompatActivity {
-    public static final String TAG = CamerasActivity.class.getSimpleName();
-    @BindView(R.id.searchView) SearchView mSearch;
-    @BindView(R.id.recyclerView) RecyclerView mRecyclerView;
+	public static final String TAG = CamerasActivity.class.getSimpleName();
+	@BindView(R.id.searchView)
+	SearchView mSearch;
+	@BindView(R.id.recyclerView)
+	RecyclerView mRecyclerView;
 	private TrackGPS mLocation;
 	private CamerasAdapter mAdapter;
 
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.cameras_activity);
-        ButterKnife.bind(this);
-        final Retrofit retrofit = RetrofitAPIClient.getClient();
-        final CameraAPI api = retrofit.create(CameraAPI.class);
-	    mLocation = new TrackGPS(this, new OnLocationChange() {
-		    @Override
-		    public void doOnLocationChange(Location location) {
+	@Override
+	protected void onCreate(Bundle savedInstanceState) {
+		super.onCreate(savedInstanceState);
+		setContentView(R.layout.cameras_activity);
+		ButterKnife.bind(this);
+		final Retrofit retrofit = RetrofitAPIClient.getClient();
+		final CameraAPI api = retrofit.create(CameraAPI.class);
+		mLocation = new TrackGPS(this, new OnLocationChange() {
+			@Override
+			public void doOnLocationChange(Location location) {
 
-		    }
-	    });
-        mAdapter = new CamerasAdapter(new ArrayList<CameraDto>());
-        mRecyclerView.setAdapter(mAdapter);
-	    mSearch.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-		    @Override
-            public boolean onQueryTextSubmit(String query) {
-                Log.d(TAG, "onQueryTextSubmit: ");
-                SharedPreferences pref = PreferenceManager.getDefaultSharedPreferences(CamerasActivity.this);
-                int dangerLevel = pref.getInt(FilterActivity.PREF_DANGER_LEVEL, 0);
-                int distance = pref.getInt(FilterActivity.PREF_DISTANCE, 5);
-	            String sortBy = pref.getString(FilterActivity.PREF_SORT_BY, "dangerLevel");
-	            double lat = mLocation.getLatitude();
-                double lng = mLocation.getLongitude();
-                RequestCameraListDto cameraListDto = new RequestCameraListDto();
-                cameraListDto.setLatitude((long) lat);
-                cameraListDto.setLongitude((long) lng);
-                cameraListDto.setDangerLevel((double) dangerLevel);
-                cameraListDto.setSearch(query);
-                cameraListDto.setSortBy(sortBy);
-                cameraListDto.setDistance(distance);
-                Call<List<CameraDto>> call = api.getAllCamerasBySearch(cameraListDto);
-                call.enqueue(new Callback<List<CameraDto>>() {
-                    @Override
-                    public void onResponse(Call<List<CameraDto>> call, Response<List<CameraDto>> response) {
-                        mAdapter.setData(response.body());
-                    }
+			}
+		});
+		mAdapter = new CamerasAdapter(new ArrayList<CameraDto>());
+		mRecyclerView.setAdapter(mAdapter);
+		mSearch.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+			@Override
+			public boolean onQueryTextSubmit(String query) {
+				Log.d(TAG, "onQueryTextSubmit: ");
+				SharedPreferences pref = PreferenceManager.getDefaultSharedPreferences(CamerasActivity.this);
+				int dangerLevel = pref.getInt(FilterActivity.PREF_DANGER_LEVEL, 0);
+				int distance = pref.getInt(FilterActivity.PREF_DISTANCE, 5);
+				String sortBy = pref.getString(FilterActivity.PREF_SORT_BY, "dangerLevel");
+				double lat = mLocation.getLatitude();
+				double lng = mLocation.getLongitude();
+				RequestCameraListDto cameraListDto = new RequestCameraListDto();
+				cameraListDto.setLatitude(lat);
+				cameraListDto.setLongitude(lng);
+				cameraListDto.setDangerLevel((double) dangerLevel);
+				cameraListDto.setSearch(query);
+				cameraListDto.setSortBy(sortBy);
+				cameraListDto.setDistance(distance);
+				Call<List<CameraDto>> call = api.getAllCamerasBySearch(cameraListDto);
+				call.enqueue(new Callback<List<CameraDto>>() {
+					@Override
+					public void onResponse(Call<List<CameraDto>> call, Response<List<CameraDto>> response) {
+						mAdapter.setData(response.body());
+					}
 
-                    @Override
-                    public void onFailure(Call<List<CameraDto>> call, Throwable throwable) {
+					@Override
+					public void onFailure(Call<List<CameraDto>> call, Throwable throwable) {
 
-                    }
-                });
-                return false;
-            }
+					}
+				});
+				return false;
+			}
 
-            @Override
-            public boolean onQueryTextChange(String newText) {
-                Log.d(TAG, "onQueryTextChange: ");
-                return false;
-            }
-	    });
-    }
+			@Override
+			public boolean onQueryTextChange(String newText) {
+				Log.d(TAG, "onQueryTextChange: ");
+				return false;
+			}
+		});
+	}
 
-    @Override
-    protected void onStop() {
-        super.onStop();
-        mLocation.stopUsingGPS();
-    }
+	@Override
+	protected void onStop() {
+		super.onStop();
+		mLocation.stopUsingGPS();
+	}
 }
